Moved Tensorflow-specific functions in tensorflow sub-package
