+++ tf_yarn/tensorflow/tasks/_independent_workers_task.py
@@ -1,22 +1,29 @@
 import logging
 
-from tf_yarn._task_commons import setup_logging
+from tf_yarn._task_commons import (
+    setup_logging, _log_sys_info, _get_experiment
+)
 setup_logging()
 
-from tf_yarn import cluster, _task_commons, _internal, Experiment, KerasExperiment
+from tf_yarn import _internal, Experiment, KerasExperiment
+from tf_yarn.tensorflow.tasks.tf_task_common import (
+    _execute_dispatched_function, _shutdown_container, _prepare_container,
+    get_task_description
+)
+from tf_yarn.tensorflow import cluster
 
 _logger = logging.getLogger(__name__)
 
 
 def main() -> None:
-    _task_commons._log_sys_info()
-    task_type, task_id = cluster.get_task_description()
+    _log_sys_info()
+    task_type, task_id = get_task_description()
     with _internal.reserve_sock_addr() as host_port:
-        client, cluster_spec, cluster_tasks = _task_commons._prepare_container(host_port)
+        client, cluster_spec, cluster_tasks = _prepare_container(host_port)
         # Variable TF_CONFIG must be set before instantiating
         # the estimator to train in a distributed way
         cluster.setup_tf_config(cluster_spec)
-        experiment = _task_commons._get_experiment(client)
+        experiment = _get_experiment(client)
         if isinstance(experiment, Experiment):
             session_config = experiment.config.session_config
         elif isinstance(experiment, KerasExperiment):
@@ -26,7 +33,7 @@ def main() -> None:
         _logger.info(f"Starting server {task_type}:{task_id}")
 
     cluster.start_tf_server(cluster_spec, session_config)
-    thread = _task_commons._execute_dispatched_function(client, experiment)
+    thread = _execute_dispatched_function(client, experiment)
 
     # "ps" tasks do not terminate by themselves. See
     # https://github.com/tensorflow/tensorflow/issues/4713.
@@ -34,7 +41,7 @@ def main() -> None:
         thread.join()
         _logger.info(f"{task_type}:{task_id} {thread.state}")
 
-    _task_commons._shutdown_container(client, cluster_tasks, session_config, thread)
+    _shutdown_container(client, cluster_tasks, session_config, thread)
 
 
 if __name__ == "__main__":
