+++ tf_yarn/tensorflow/tasks/_tensorboard_task.py
@@ -4,13 +4,12 @@ import os
 import skein
 
 from tf_yarn import Experiment, KerasExperiment
-from tf_yarn._task_commons import setup_logging
+from tf_yarn._task_commons import setup_logging, get_task, get_task_description
 setup_logging()
 
 from tf_yarn import (
     _task_commons,
     _internal,
-    cluster,
     event
 )
 from tf_yarn.tensorflow import tensorboard
@@ -20,8 +19,8 @@ _logger = logging.getLogger(__name__)
 
 def main() -> None:
     _task_commons._log_sys_info()
-    task_type, task_id = cluster.get_task_description()
-    task = cluster.get_task()
+    task_type, task_id = get_task_description()
+    task = get_task()
     client = skein.ApplicationClient.from_current()
 
     _task_commons._setup_container_logs(client)
