+++ tf_yarn/tensorflow/tasks/gloo_allred_task.py
@@ -15,9 +15,11 @@ except (ModuleNotFoundError, ImportError) as e:
     warnings.warn("Horovod is not installed. See README for instructions to install it")
     raise e
 
-from tf_yarn import event, _task_commons, cluster, Experiment, KerasExperiment
+from tf_yarn import event, _task_commons, Experiment, KerasExperiment
 from tf_yarn.tensorflow.tasks.evaluator_task import evaluator_fn
-from tf_yarn._task_commons import setup_logging
+from tf_yarn._task_commons import (
+    setup_logging, get_task, is_chief, get_task_type, get_task_description
+)
 setup_logging()
 
 logger = logging.getLogger(__name__)
@@ -53,7 +55,7 @@ def _worker_fn(client, task, net_if):
     experiment = _task_commons._get_experiment(client)
 
     if isinstance(experiment, Experiment):
-        if not cluster.is_chief(cluster.get_task_type(task)):
+        if not is_chief(get_task_type(task)):
             # Overwrite config to do nothing but training to improve training speed
             experiment.estimator._model_dir = "."
             new_config = experiment.estimator.config.replace(
@@ -71,7 +73,7 @@ def _worker_fn(client, task, net_if):
             hooks=experiment.train_spec.hooks,
             max_steps=experiment.train_spec.max_steps)
     elif isinstance(experiment, KerasExperiment):
-        if not cluster.is_chief(cluster.get_task_type(task)):
+        if not is_chief(get_task_type(task)):
             if experiment.train_params['callbacks'] is not None:
                 callbacks_to_keep = []
                 for callback in experiment.train_params['callbacks']:
@@ -108,19 +110,19 @@ def _driver_fn(client, net_if):
             {host.rank},{host.size},{host.local_rank},\
             {host.local_size},{host.cross_rank},{host.cross_size}\
             """
-        event.broadcast(client, f"{cluster.get_task()}/{host.hostname}", host_info)
+        event.broadcast(client, f"{get_task()}/{host.hostname}", host_info)
 
     global_rendezv = RendezvousServer(verbose=1)
     global_rendezv_port = global_rendezv.start_server()
     global_rendezv.httpd.init(host_alloc_plan)
-    event.broadcast(client, f"{cluster.get_task()}/sock_addr", f"{net_if[1]}:{global_rendezv_port}")
+    event.broadcast(client, f"{get_task()}/sock_addr", f"{net_if[1]}:{global_rendezv_port}")
     return global_rendezv.listen_thread
 
 
 def main():
     client = skein.ApplicationClient.from_current()
-    task_type, task_id = cluster.get_task_description()
-    task = cluster.get_task()
+    task_type, task_id = get_task_description()
+    task = get_task()
     event.init_event(client, task, "127.0.0.1:0")
     _task_commons._setup_container_logs(client)
     net_if = get_net_if()
